% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_jobs}
\alias{get_jobs}
\title{Return the jobs stored in the storage table}
\usage{
get_jobs(n = 10, raw = FALSE)
}
\arguments{
\item{n}{the number of jobs to return. Default: `10`}

\item{raw}{flag indicating if the `tbl` of the
database connection is returned or if the data is
sorted by `timestamp` and `collect()` is called.
Default: `FALSE`}
}
\description{
Return the jobs stored in the storage table
}
