% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{job_status}
\alias{job_status}
\title{The job status using a defined vocabulary}
\usage{
job_status(job)
}
\value{
Either "done", "running", "failed", or "not_stated"
}
\description{
The job status using a defined vocabulary
}
