% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lapply_to_dependencies}
\alias{lapply_to_dependencies}
\title{Apply function to all dependencies of a job and return a list}
\usage{
lapply_to_dependencies(x, FUN)
}
\arguments{
\item{x}{a job or a list of jobs}

\item{FUN}{the function that is called with a job object}
}
\description{
Apply function to all dependencies of a job and return a list
}
