% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{output_log_file_path}
\alias{output_log_file_path}
\alias{get_output_log}
\alias{show_output_log}
\title{Get the output path for the log file of a job}
\usage{
output_log_file_path(job)

get_output_log(job)

show_output_log(job, print = FALSE)
}
\arguments{
\item{job}{a job object}
}
\description{
Get the output path for the log file of a job
}
