% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{result_file_path}
\alias{result_file_path}
\alias{show_result_file}
\title{Get the output path for the result file of a job}
\usage{
result_file_path(job)

show_result_file(job, print = FALSE)
}
\arguments{
\item{job}{a job object}
}
\description{
Get the output path for the result file of a job
}
