% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{slurm_job_status}
\alias{slurm_job_status}
\title{Slurm job status for a job}
\usage{
slurm_job_status(job)
}
\description{
Typically either "PENDING", "RUNNING", "FAILED", "CANCELLED", or "COMPLETED"
but there are several more options (see https://slurm.schedmd.com/sacct.html#SECTION_JOB-STATE-CODES)
}
