% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stats_file_path}
\alias{stats_file_path}
\alias{show_stats_file}
\title{Get the output path for the stats file of a job}
\usage{
stats_file_path(job)

show_stats_file(job, print = FALSE)
}
\arguments{
\item{job}{a job object}
}
\description{
Get the output path for the stats file of a job
}
